Many modifications per BNL meeting. All Specific Ntypes changed.
Normative Types.

Q: Right now, NTHistogram and NTVariantArray don't have a field named "value". Might
want to coerce them to have, so a trivial deserializer can always pick out a meaningful value.
Also this would be the value made available through CA.

* Removed most explicit text to do with Ntype self-identificaition. This leaves that
deliberately undefined so we have to define it.

* Tightened syntax and grammar for descriptions of fields and types, in particular to
distinguish non-terminals from terminals, and to distinguish the Normative Type data
definition from its implementation in a pvData binding. Maybe we want to be even more explicit.

  In fact added css style rules to help format grammar elements appropriately. It's cool,
check it out. When Ben complains that I've used the opposite of the BNF conventions for terminality,
I can change the rendering of all symbol types just by changing the css!

* Flattened NT Fields that used structures within structures. Only programmers could love that.
And the syntax defied above it didn't allow for it anyway.

* Introduced NTVariantArray
     Note, NTVariantArray isn't defined by first defining a variant_t and then making an array of
     them, because then each element of the array could have different type. We could in addition
     to NTVariantArray define a variant_t and hence variant_t array, but that would have different
     use cases.

* Replaced NTImage's use of "variant" with NTVariantArray

* Replaced NTArray by name NTMatrix, and tightened up required interpretation of "dim"

(Now NTMatrix covers what you'd use NTArray for. NTVAriantArray would be used when
you want to use the same structure instance, but change the value type).

* Changed definition of NTNameValue so that it defines a single name and value pair, rather
than congruent arrays (that former functionality is given by NTMatrix and NTTable).
And changed from simple scalar to scalar array, so can transport an array of n-tuples (of
possibly different length).

* Changed NTTimeDomainArray, so only numeric scalars may be used.

* Re NTFrequencydomainArray changed field name firstFrequency to baseFrequency. Changed
definition from using firstFrequency and deltaFrequency to using a congruent array
of frequencies. This is necessary right? Don't think it could be done with just 2 values.

* NTHistogram.
  Many changes and additions including interpretation.

* NTStatistic
  renamed NTAggregate
  A lot of stuff

* NTTable. Changed from transporting arrays of any normative type, to only
arrays of scalar. This should make conforming implementations much much simpler,
and removes questions of recursion. It also means an NTTable can always be intuitively
rendered in 2d as recognisably a table.
	If we want a more complex structure for whole datasets, then let's add "NTDataSet"
and make it's implementation in some sense optional. Maybe homogenous array and
heterogenous array, as in AIDA.

* NTImage. Changed grammatical expression to use non-terminals for the 3 parts. Also refer
explicitly to use of an embedded NTVariantArray. Made array part use only numeric scalars.

Q: Why are the array type NT Filed's described as "anyName" allowed? I would have thought if one
is going to use the structure name to identify a Ntype, then a structure field of an NType
would also need to self identify in that way.

** have to decide whether Ntypes cover processing or only interop, since depth of normative
definition of NTMatrix is very dependent on that. The layout of elements in memory - row
or column major order, would have to be defined if NType covers processing.